requirements.txt
django==3.0.4
djangorestframework==3.11.0
requests